Typecast columns when creating model objects from JSON in the json_serializer plugin
This changes the logic to try the setter method before just modifying
the values hash for the model's columns.  The values will still be
modified for columns that don't have an available setter method, or
where the setter method is restricted (e.g. primary key columns).

Since the setter methods are being used for intake, the output
code has been changed to call the getter methods instead of directly
getting the value from the values hash.