Allow the user to control how the connection pool deals with attempts to access shards that aren't configured
Connection pools now take a :servers_hash option that will be used
as the hash for the servers.  The default for this option is
Hash.new(:default), so that attempts to access unconfigured shards
end up using the default shard (the same as Sequel's historical
behavior).

You can use something like Hash.new(:read_only) to automatically
use the :read_only shard.  If attempts to access an unconfigured
shard should be considered an error, you can use a hash with
a default proc that raises an error, allowing easier detection
of this case.