Mark the generic file type as pending for jdbc and postgres
There isn't a 9.0 version of the jdbc-postgres gem, and when
you run the 8.4 version against a 9.0 server, the test fails
because it doesn't handle the bytea unescaping properly. You
can set a function on the server to use the previous
escaping to get around this, but it's best just to mark the
test as pending until a 9.0 version of the jdbc-postgres
gem is released.