Use methods instead of procs for type translation in the postgres adapter
This adds a TYPE_TRANSLATOR object with real methods, and creates
method objects from it, instead of the previous code which used
lambda procs.  This code is up to 15% faster according to my
testing, which makes sense because procs are slower than methods,
since procs are closures and methods aren't.

This also switches the use_iso_date_format setter for the Postgres
module to change which method is used, instead of checking the
value of use_iso_date_format inside the method.