Use methods instead of procs for type translation in the mysql adapter
Similar idea with simpler execution than the changes to the mysql
and postgres adapters. One small change is that instead of using
Integer() and Float(), to_i and to_f are used.

The SQLITE_TYPE_PROCS constant is being eliminated, as are the
UNIX_EPOCH_TIME_FORMAT and FALSE_VALUES constants.