Allow reflecting on validations in the validation_class_methods plugin
This is a super simple implementation that allows you to determine
which validations were used on the model.  It's only added to the
validation_class_methods plugin, as the validation_helpers plugin
is instance level and not class level.

Basically, this just copies the options given in each validations
call and makes them available.  They are stored in a simple
hash in the model class that you can get by calling
validation_reflections.  The hash is keyed by column symbol, with
values that are an array of two element arrays, where the first
element is the validation type symbol and the second is the
options hash.