Use varbinary(max) instead of image for MSSQL blob type
image is currently deprecated, and with the current code, if you
use a blob in a prepared statement and compare it to an image
column, you get an error, because it assumes the blob is a
varbinary and the two aren't compatible.

This breaks backwards compatibility, but I think it's the
right step to take.