Make class_table_inheritance plugin handle subclass associations better (Fixes #320)
Previously, if you loaded a dataset for a model that uses
class_table_inheritance, and then on one of the instances
you called an association method for an association that
existed in the class of the instance but not in the main
class, a NoMethodError was raised.

This adds an Sequel::UndefinedAssociation exception class
that can be raised by the eager loading code when it sees
an attempt to use an undefined association.  The
tactical_eager_loading plugin (which the class_table_inheritance
plugin loads implicitly) will recognize and ignore this error
code, since a failed eager loading attempt will still be followed
by a regular lazy loading attempt.