Use a better error message if a validates_max_length validation is applied to a nil value (Fixes #322)
This patch changes the default error message used for this case to
be "is not present". The previous message was misleading, as it
said that the value was too long.  You can configure this message
as well using the :nil_message option.

Both the validation_helpers and the validation_class_methods
plugins were updated to have roughly the same API.  For
validation_class_methods, this will affect code such as:

  validates_length_of :foo, :maximum=>bar

I don't really like ugly special casing like this, but I can't
think of a cleaner way that keeps backward compatibility and
fixes the bogus error message.