Add tinytds adapter, the best way to connect to MSSQL from a C based ruby running on *nix
Before, the only real way to connect to MSSQL from a C based ruby
on *nix was to use the ODBC adapter with unixodbc and freetds.
I've heard it's a big pain to set up, and never attempted to do
so myself.  Fortunately, tiny_tds was released recently and makes
it simple.

Thanks to metaskills from the tiny_tds project for making changes
to tiny_tds to better support Sequel.