Combine -r and -L load commands to execute in order
Previously, -r files would be loaded before the DB constant was
established.  Now, it holds the information in the same data
structure as the -L option, which enables it to be loaded in
the same order.  So if you specify -L before -r, the -L dir will
be loaded before the -r dir is required, and vice versa.