Make using NULL/NOT NULL, DEFAULT, and UNIQUE column options work correctly on H2 and possibly Oracle
Before, Sequel had a default order that it used for column
definitions.   The columns options used UNIQUE, then NULL/NOT NULL,
then DEFAULT.  However, that is not valid SQL92.  In SQL92, DEFAULT
always comes before NULL/NOT NULL and UNIQUE (which are constraints).

While it may not be strictly necessary, I decided to break up the
column_definition_sql method into smaller parts with a configurable
order, similar to how dataset queries are built.  While MySQL appears
to work with the new default order, based on the MySQL documentation,
NULL/NOT NULL comes before default, so I decided to change the order
on MySQL (maybe older versions of MySQL enforce it?).

This passes on SQLite, MySQL, PostgreSQL, MSSQL, and H2.  I'd greatly
appreciate testing on other databases.  Part of the reason for this
change is that Oracle also requires that the SQL92 behavior of
defaults being specified before constraints.