Use the correct class when loading instances for descendents of model classes that use single table inheritance
Before, only the model class that loaded the
single_table_inheritance plugin had the logic to return instances
of subclasses.  This copies that logic into subclasses, so that if
you load all instances of a subclass, it can return instances of
subsubclasses.  Thanks to Max Aller on sequel-talk for pointing
out this bug.