Add Dataset/Database#mssql_unicode_strings accessor when connecting to MSSQL to control string literalization
This is just a cleanup of the previous commit.

By default, Sequel uses N'' for string literalization on MSSQL.
However, it can be faster to use just '', though that does not
work correctly in all cases.  This allows the user to use one
of the following to change the string literalzation to use '':

  DB.mssql_unicode_strings = false # default for datasets
  dataset.mssql_unicode_strings = false # just this dataset

You can also use the :mssql_unicode_strings option when connecting
to the database.

For more details on the difference between N'' and '', see
http://support.microsoft.com/kb/239530.