Modify how eager_graph callback is handled
This reverts the changes to dataset/graph.rb.  Dataset#graph should
not be coupled at all to eager loading.  This renames the
:eager_block option to :callback and handles it inside the
:eager_grapher procs instead of pushing the changes down to
Dataset#graph.

Along with the :eager_block parts of dataset/graph.rb, I'm also
reverting the use of dataset.opts[:select] to get the dataset's
columns.  I'm guessing this was done for performance reasons to
avoid a query, but there are better ways to solve it.  The
problem with using dataset.opts[:select] is that if you have
any columns that aren't symbols or any symbol columns like
:table__column, :column___alias, or :table__column___alias,
it won't give you the correct result.  I'll probably add an
extension that ships with Sequel that will do Dataset#columns
by introspection if possible instead of just sending a database
query.

Because of the mocking in the specs, I currently need to manually
set the columns on the graphed datasets to get the specs to patch.