Support cascading of eager loading using an intermediate proc for both eager and eager_graph
Similar functionality was already available with eager by calling
eager again inside the proc.  However, that method is not
compatible with eager_graph, since the proc is called with the
dataset to graph into the current dataset.

For example, if you have this code:

  Artist.eager(:albums=>:tracks)

And you only want to load albums where the id > 100, you do:

  Artist.eager(:albums=>{proc{|ds| ds.where{id > 100}}=>:tracks})

Basically, you replace the cascading models with a single entry
hash that has a proc callback key and the cascading model values.
This syntax works for both eager and eager_graph.