Support dynamic callbacks to customize regular association loading at query time
This brings the dynamic callback feature recently added to eager and
eager_graph and applies it to regular association loading.  You can
now apply a dynamic callback for regular associations by passing
a block to the association method:

  artist.albums{|ds| ds.where{year > 1990}}

On ruby 1.8.6, that isn't supported, as 1.8.6 doesn't allow blocks
to take block arguments.  So there you have a pass the proc as
an argument:

  artist.albums(proc{|ds| ds.where{year > 1990}})

This breaks backwards compatibility in the private API for the
load_associated_objects and _load_associated_objects methods.
They now take 2 arguments instead of 1, the new argument being
a dynamic options hash.

While here, fix some breakage on ruby 1.9 for multiple
assignment inside a conditional.