Add a columns_introspection extension that attempts to skip database queries by introspecting selected columns
Sequel's previous behavior has been to do a database query to
retrieve the columns for a dataset unless the dataset has already
cached the columns.  This commit adds an extension that makes
Sequel attempt to introspect the selected columns to guess at
the columns that would be returned.  This should work in most
cases, but as there is no guarantee that Sequel will guess
correctly, this is not being done by default.

To use this, after loading the extension, you can extend any
dataset with Sequel::ColumnIntrospection.  If you want to use
this for all datasets, run:

  Sequel::Dataset.introspect_all_columns

This adds some hooks to the specs so that all of Sequel's specs
can be run with this extension.  To do so, define the
SEQUEL_COLUMNS_INTROSPECTION environment variable when running
the specs.