Don't call insert_select when saving a model that doesn't select all columns of the table
insert_select is an optimization used by some adapters that does
the insert and the refresh of the row in the same query.  However,
it always returns all columns, and the API would have to change
to select specific columns.  Since this is just an optimization,
it's better just to skip it if the model doesn't select all
columns.

The reason this is being done is that one of the main reasons
to restrict the selected columns for a model is because one of
the columns that isn't selected is very large.  If insert_select
is used, it causes the very large column to be loaded when that
probably isn't desired, which would reduce performance instead
of increase it.