Split informix adapter into shared/specific parts, add JDBC informix subadapter
This is a fairly straightforward split of the informix adapter
into shared specific parts, for the purpose of adding a JDBC
informix subadapter so that informix is reachable from JRuby.
Tested by Christian Michon on the Google Group.  I haven't
tested the native informix adapter (no access to an informix
installation), but it looks unlikely to break things.