If Thread#kill is called on a thread with an open transaction, roll the transaction back on ruby 1.8 and rubinius
Ruby 1.9 and JRuby don't handle Thread#kill the same way, and it's
not possible to rely on them to fix this issue.  In general,
Thread#kill should never be used, but since it is called by
default when a ruby process terminates (if it has threads open),
it's something I think should be handled by default.

Idea taken from coderrr's blog post, see:
http://coderrr.wordpress.com/2011/05/03/beware-of-threadkill-or-your-activerecord-transactions-are-in-danger-of-being-partially-committed

For some reason, on rubinius the DataObjects SQLite subadapter
and the tinytds adapter both hang in the Thread#kill tests, so
skip the tests for those adapters.