Remove all uses of sleep in the connection pool specs, using queues instead
Sleeping is inherently racy.  There have been quite a few commits
in the past to adjust the sleep time, and the connection pool
specs were the slowest ones because of all of the sleeping.

This commit replaces all sleep calls with the use of queues
(usually 2) in order to synchronize the threaded code.

Most of the changes are fairly straightforward, though I did
remove some calls to Thread#raise, since those didn't work
correctly in JRuby, and weren't really necessary for the
specs.

This reduces spec runtime for the connection pool spec from 2.3 to
0.35 seconds on ruby 1.8, and from 2.1 to 0.21 seconds on ruby 1.9
on my main test box.