Add Database#views for getting an array of symbols of view names for the database
Database#views operates very similar to Database#tables, just
returning view names instead of table names.  In all the adapters,
this turned out to be as simple as parametizing the existing table
method into a new method and having tables and views call it.

While here, fix an issue with the MySQL support returning view
names in Database#tables by using SHOW FULL TABLES.

Also, reorder some methods so that they are in alphabetical order.

In the specs, actually check that tables includes a table and
views includes a view, and that tables doesn't include views
and views doesn't include tables.