Make create_table? use IF NOT EXISTS instead of using SELECT to determine existence, if supported
This is in response to an official request by Wayne Seguin (who used
to contribute to Sequel regularly back in late 2007-early 2008).

The CREATE TABLE IF NOT EXISTS syntax is supported by MySQL and H2,
and SQLite 3.3.0+.  Those are the only databases that Sequel enables
this support on.  If other databases implement it, please let me
know.