Add Sequel::Postgres::PG_NAMED_TYPES hash for extensions to register type conversions for non-standard types
This makes the conversion procs modifiable per Database object,
by duping the PG_TYPES hash.  Because non-built-in types can
have varying oids per database, the PG_NAMED_TYPES hash is
introduced allowing extensions to associate types by type name.
Then, when the first connection is made, it parses the system
type catalog.  For any type names that are in PG_NAMED_TYPES,
if there is not an existing entry for the type oid, the callable
value from PG_NAMED_TYPES is added to the database's conversion
procs by oid.  This way, when fetching columns, types are still
looked up by oid instead of by name.

Note that if you are using the sharding support and want to use
custom types and type conversion procs, you need to make sure that
all shards have the same oids for the type names.

No tests for this support yet, but they should come soon.