Add support for filtering and excluding by multiple associations
This allows you to use the following types of filters:

  Album.filter(:artist=>[artist1, artist2])
  Artist.exclude(:album=>[album1, album2])

When multiple associated objects are used in an association filter,
the filter matches if any of them are true, similar to using:

  Album.filter(:artist_id=>[artist1.id, artist2.id])

For exclusions, the filter matches if the album doesn't match
any of them.  It's similar to:

  Artist.exclude(:id=>[album1.artist_id, album2.artist_id]).
   or(:id=>NULL)

Composite key associations are fully supported both for filters
and exclusions.

There are a lot of corner cases to be dealt with, and this
commit attempts to do so.  Corner cases like the value of
the keys being NULL, having arrays of model objects provided
where some objects have NULL keys and others do not, and cases
where composite keys are used when some of the composite keys
are NULL and others aren't.  Some of the corner cases would not
be handled correctly previously, others are just related to this
new multiple object support.

This is not fully ready for production as I still need to write
integration tests for this as well as adding support for
many_through_many associations.