Fix literalization of booleans on SQLite
SQLite uses 0/1 for booleans, instead of 't' and 'f'.  Unfortunately,
just changing the literalizations via literal_true and literal_false
breaks things, because SQLite then can't tell the difference between
an integer value and a boolean value.  This is an ugly hack to try
to get filters to use 0/1 while still using 't' and 'f' for
compatibility elsewhere.

While here, add an integration test to see if any other adapters
are similarly broken.