Add support for filtering and excluding by association datasets
This adds the ability to filter and exclude by association
datasets the same way you filter and exclude by associations or
multiple associations.  For example, this would get you the
albums by all artists whose name starts with A:

  Album.filter(:artist=>Artist.filter(:name.like('A%')))

Just like the previously committed support, this new support
works with composite key associations and handles NULL values
properly.

This could break backwards compatibility for some apps, though you
would need to be using an association name symbol in a filter with
a dataset whose model is the same as the associated class (or a
subclass of that class).  As associations almost never have the
same names as columns, you'd either be aliasing or joining to
another table to hit this.  The probability of production code
actually breaking because of this is low enough I'm willing to
take the risk.  If the stars align poorly for you and you hit a
bug because of this, you can fix it by changing the association
name symbol to an SQL::Identifier by calling #identifier on it,
or using a literal string instead.