Only typecast bit fields to booleans on MSSQL, the MySQL bit type is a bitfield, not a boolean
This was pointed out by axic on GitHub.  It's possible that MySQL
bitfields will be supported in some future version, but it'll
probably require a custom type (the equivalent of Sequel::SQL::Blob
but for bitfields).

If any other databases use bit as a boolean value, this will
break some backwards compatibility.  Please let me know if your
database is affected.