Add Dataset#with_pk for model datasets, for finding first record with matching primary key value
This is being added to make it easier to do a primary key lookup.
Sequel has supported Model.[] for lookups by primary key, but
if you wanted to add a filter to that, you had to give the
primary key explicitly:

  Model.filter(:foo=>'bar')[:id=>1]

With the addition of the with_pk method, you can do:

  Model.filter(:foo=>'bar').with_pk(1)

Which doesn't require you specify the primary key.

Also, Dataset#[] has been overridden to call with_pk if given a
single integer value, so you can do:

  Model.filter(:foo=>'bar')[1]

Unfortunately, due to backwards compatibility, that won't work
with strings or composite primary keys, so you will have to use
with_pk for those.  For composite primary keys, you have to use
an array:

  Model.filter(:foo=>'bar').with_pk([1, 2])