Add Dataset#unbind for unbinding values from a dataset, for use with creating prepared statements
This adds a new ASTTranformer subclass that does the work, by
recognizing some types of ComplexExpressions, extracting
bound variables and substituting them with placeholders.  The
statements can then be prepared and executed (or just called)
with the extracted bound variables.

Also, fix a small issue with using double underscores for
placeholder values in the SQLite native support for argument
mapping.