Move Dataset#with_pk prepared statement use to prepared_statements_with_pk plugin
This functionality now uses Dataset#unbind to create a more generic
dataset, which lowers the risk of an unbounded number of prepared
statements being create, but it does not eliminate the risk.  This
plugin should only be used by those who are sure that their
usage of Dataset#with_pk cannot result in a denial of service
attack.