Add prepared_statements_safe plugin, for making prepared statement use with models more safe
This new plugin doesn't use prepared statements at all, but it's
designed to be used with (and requires) the prepared_statements
plugin.

The basic security issue with using prepared statements implicitly
with Sequel is that Sequel by default only uses uses the currently
present columns when insert (some subset of the table's columns),
and by default when updating only saves the changed columns.

For prepared statements to be used, each set of columns in the
insert and update statements needs to have its own prepared
statement.  If you have a table with 1 primary key column and
4 other columns, you can have up to 2^4 = 16 prepared statements
created, one for each subset of the 4 columns.  If you have 1
primary key column and 20 other columns, there are over a million
subsets, and you would assuredly hit your database limit for
prepared statements (a denial of service attack).

The fix for this is to use every column possible when inserting
and updating.  For updating, this is simple, as you just save
all columns.  For inserting, this isn't always possible, as
you can't necessarily insert a correct default value, as it
could depend on a database function.  So for NULL defaults and
defaults that Sequel can parse, Sequel will add those columns
to the insert statement.