Treat not calling super/yielding in an around hook like a before hook returning false
Because around hooks can now raise exceptions similar to before
hooks, rename the exception to HookFailed, but still assign it
to the BeforeHookFailed constant as well for backwards
compatibility.

With this change and the previous one, around hooks are now very
similar to ActiveSupport ones in terms of behavior, though they
use instance methods instead of blocks.