Add _valid? private model method so that valid? can always return boolean
valid? should always return a boolean even if a validation hook
fails.  However, when save validates, a HookFailure exception
should be raised instead.  The _valid? method takes a parameter
for whether to raise or return false for hook failures.

It's no longer safe to override valid?, which I was doing in a few
specs.  You should only be overriding validate to add errors.