Support :cascade option to Database#drop_table and #drop_view for using CASCADE
This makes drop_table and drop_view accept an options hash as the
final argument.  Currently they only support the :cascade=>true
option for turning on CASCADE.

No longer use CASCADE unilaterally when dropping tables on
PostgreSQL.  This breaks backwards compatibility, but is in my
mind the sanest choice.  CASCADE is a dangerous option, and
using it implicitly is a bad idea.  The usage of CASCADE by
default dates back to the original import of Sequel code into
the subversion repository on February 28, 2007.

This adds the private drop_view_sql method, and makes drop_table_sql
take two arguments.

This also moves some schema related specs to the schema_spec file.