Work with tinytds 0.4.5
This removes the copying of the :host option to the :dataserver
option when using the tinytds adapter.  That worked fine for previous
versions of tiny_tds, but no longer works with 0.4.5.

This breaks backwards compatibility with previous connection urls
where the freetds.conf name was specified as the :host.  To use
an entry in the freetds.conf file, you must now provide the
:dataserver option.

Unfortunately, tinytds 0.4.5 does not work with FreeTDS 0.91rc1, you
have to use FreeTDS 0.91rc2 or an older version of FreeTDS such as
0.82 (and older versions require the use of the freetds.conf file).

Since 0.4.5 is now required, switch to using active? instead of
closed? and dead? to detect disconnects.