Recognize disconnections when issuing BEGIN/ROLLBACK/COMMIT statements (Fixes #368)
Add Database#disconnect_error? private method that raise_error calls
to determine whether to raise a DatabaseError or
DatabaseDisconnectError.  In some cases, the disconnect determination
depends on the exception, but in other cases it depends on the
connection, so pass the connection in a :conn option and
use that if available.

This also cleans up the handling of regular disconnects, since in
most cases you don't need to specify the :disconnect option manually
(though that still works).