Make create_table! only attempt to drop the table if it already exists
Before, it dropped the table unconditionally, then attempted to
create it.  This can hide bugs, since there are other reasons a
drop would fail besides nonexistence of the table.  The most
common cause would be that there are other tables that still
reference the table. Because the previous implementation did
rescue nil, the error got swallowed and you then got an error
during the CREATE TABLE statement saying the table already
exists.