Don't treat strings with leading zeros as octal format in the default typecasting
It's much more likely that someone will use '013' to mean 13
instead of 11, and raising an error for '08' is probably a bad
thing.  Still respect the 0x prefix for hexadecimal.

If anyone is relying on the old behavior, I'll be happy to add
an extension that restores it (similar to how the
looser_typecasting extension uses to_i and to_f).