Use hanna-nouveau instead of hanna as the RDoc template
mislav's original hanna only works on RDoc 2.3.  It's pretty much a dead
end development wise.  hanna-nouveau is a fork that works with modern
versions of RDoc (currently 3.8).

This uses hanna-noveau if a modern version of RDoc is used, and
should still retain backwards compatibility with really old versions
of RDoc.  I'm not sure if it works with newer RDoc if you don't have
hanna-nouveau installed, though.