Major speedup in dataset creation
This takes the setting of the @quote_identifiers,
@identifier_input_method, @identifier_output_method, and @row_proc
out of Dataset#initialize and delays their initialization until
they are needed.  This makes dataset creation about 2.5x faster on
1.8 and 4.4x faster on 1.9, and dataset cloning about 40% faster
on 1.8 and 20% faster on 1..

The tradeoff is about a 33% decrease in performance of
input_identifier and output_identifier, which are not usually
called directly.  Overall performance decrease during literalization
is very small.  This is likely to be an overall performance benefit
for almost all Sequel code.