Disallow eager loading via eager of many_*_many associations with :eager_graph option
This combination (eagerly loading via eager, :many_*_many
association, :eager_graph option), was supposedly supported
earlier, and even had specs (but no integration tests).  Well, the
specs had invalid SQL in them (referenced tables only used in
subqueries), so they could never work on a real database.

Since this apparently never worked correctly, I'm just removing
the feature.  Fixing it would complicate the EagerGraphLoader,
and probably make it slower, and as apparently nobody is using
the feature, there's no point to having it.  I'm open to adding the
feature back at a later point if someone submits a patch that uses
the correct SQL (not hard), fixes the the EagerGraphLoader to
handle it (also not hard, most of the code needed is being removed
in this commit), and includes a full integration test suite.

This feature was written before Sequel used integration tests, which
is why there were no integration tests for it.  I didn't actually
catch the problem with the SQL until writing the integration test
for it.

I don't think this causes any backwards compatibility issues, other
than you may get a Sequel::Error raised instead of a
Sequel::DatabaseError.