Fix connection in mysql adapter with :port=>"1234"
This can happen if the connection information is stored in yaml.
When a connection string is used, the URI library turns the port
into an integer, but that does not happen if an options hash is
used directly, and parsing yaml to get an options hash will
generally result in string values.