Model#exists? now returns false without issuing a query for new model objects
Previously, it could return true if you set the primary key field
in advance and the value already existed in the database.  However,
attempting to insert the object would cause an error due to the
duplicate primary key.

This does break backwards compatibility, but nobody should have
been relying on the old behavior.