Handle allowed mass assignment methods correctly when including modules in classes or extending instances with modules
Before, including a module in a class would not clear the setter
cache, so if the module included a setter method you wanted to
allow, it may have been ignored.  Likewise, if you extended the
object with a module containing a setter method, that didn't set
the flag to ignore the model class's cache.

In both cases, the handling is unconditional, it doesn't check
that the module includes a setter method.  That may change in the
future, but shouldn't cause any problems now.