Add Model.dataset_module for simplifying many def_dataset_method calls
This can be used to clean up a bunch of def_dataset_method calls:

  Artist.def_dataset_method(:foo) do
    :bar
  end
  Artist.def_dataset_method(:baz) do |x|
    x + 2
  end

becomes:

  Artist.dataset_module do
    def foo
      :bar
    end
    def baz(x)
      x + 2
    end
  end

This also makes the dataset methods significantly faster as
the methods are no longer closures.  So even for only a single
dataset method, it makes sense in performance sensitive situations.
Another benefit is that since the methods are defined normally
instead of using define_method with a block, argument handling
is stricter.