Add support for native prepared statements and stored procedures to the mysql2 adapter
The mysql and mysql2 adapters are very similar, so this extracts
the existing support from the mysql adapter into a shared adapter
that both the mysql and mysql2 adapter can use.

mysql2 doesn't appear to support stored procedures that return
results, but it does support other types of prepared statements.
I think this could be fixed by setting flags when connecting,
but I'm not sure that Mysql2 supports the same flags that
Mysql does (it has CLIENT_PS_MULTI_RESULTS but not
CLIENT_MULTI_RESULTS).