Significant changes to how the Sequel DSL handles corner cases
The main feature added by this commit is changing the + operator
to do string concatenation instead of addition if given a string,
and the & and | operators to do & and | instead of AND and OR
if given a number.

  (:x + 'a')
  # Before: (x + 'a')
  #  After: (x || 'a')
  (:x & 1)
  # Before: (x AND 1)
  #  After: (x & 1)

Those changes are only made when Sequel doesn't have an indication
of the SQL type.  For example:

  (:x + 1 + 'a')
  # (x + 1 + 'a')
  (:x & :y & 1)
  # (x AND y AND 1)

This is because x + 1 should result in a numeric type, and x AND y
should result in a boolean type.

This removes a lot of the defense code in the DSL designed to ensure
that you can't create types that don't make logical sense, such as
performing AND or OR operations on numbers or strings or greater or
less than opreations on booleans.  However, many databases allow such
operations, and Sequel should be a tool to allow you to interact with
the database, not an arbiter of what is allowed or not allowed.

This could break backwards compatibility if you are doing :x + 'a'
or :x & 1.  Anyone doing that needs to switch to :x.sql_number + 'a'
or :x.sql_boolean & 1.

This changes the bitwise operators to combine multiple
expressions with the same operation, eliminating unnecessary
parantheses.

This switches the metaprogramming used in the DSL to define many
similar methods from using define_method to using module_eval,
which should speed up such method calls by about 50% on both 1.8
and 1.9.

This adds sql_{boolean,number,string} methods to complex expressions,
allowing to return a new expression with the same operator and
arguments but with different behavior, allowing you to do things
such as:

  (x + 1).sql_string + 'a'
  # ((x + 1) || 'a')