Add Database#use_timestamp_timezones= accessor on SQLite
This allows you to turn off the use of timezones in timestamps
by setting the value to false.  This is necessary if you want
you want to use the SQLite datetime functions.

Note that this setting does not affect the current database
content.  To convert old databases to the new format, you'll
have to resave all rows that have timestamps.