Association :after_load hooks in lazy loading are now called after the associated objects have been cached
Before, it was not possible to change the value cached for the
association or the value returned, as the after_load hooks were
called before the caching was done.  This delays calling the
after_load hooks until after the caching is done.

This makes :after_load for the lazy load case have the same
ability that :after_load has if you use eager to load the
objects.