Allow Model.set_dataset to accept Sequel::LiteralString arguments as table names
I originally was going to allow all strings, but the fact that you
can use plain strings as an argument to Dataset#from is not something
that I think should be relied on.  For one thing, it breaks if you
later use the table name as an argument to #literal.