Add Dataset#supports_cte_in_subqueries? for checking whether database supports WITH in subqueries
This extracts the code that was used in the shared MSSQL adapter,
and uses it by default.  Since PostgreSQL is the only database that
appears to support CTEs in subqueries, it is the only database where
such support is enabled.

Note that this CTE copying from subqueries to the main query is
currently only used in join_table.  In the future, support will
probably be added to from, union, intersect, except, and other
methods that take a dataset.  Subqueries can appear other places
(almost anywhere in the query), but because of how Sequel works,
supporting them everywhere would require preprocessing the entire
dataset when literalizing it, which is a performance hit that
probably isn't worth taking.