Significant changes to get the ibmdb adapter passing the test suite
This makes the ibmdb adapter pass all but 6 tests (which are marked
as pending) using DB2 9.7.  Still various cleanup to do in the
adapter, but this is a good baseline to make sure that no regressions
are introduced.

Rename the spec file to db2_spec, since specs are generally for
databases and not adapters.